# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Young Tech Challenge - Dokumentacja API",
        "description": "API do automatycznego przetwarzania faktur PDF przy użyciu OCR oraz sztucznej inteligencji (Groq AI) w projekcie Laravel.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Lokalny Serwer API"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Autoryzacja"
                ],
                "summary": "Rejestracja nowego użytkownika",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jan Kowalski"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "haslo123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Użytkownik pomyślnie zarejestrowany."
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autoryzacja"
                ],
                "summary": "Logowanie użytkownika",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "haslo123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pomyślnie zalogowano, zwrócono token."
                    },
                    "401": {
                        "description": "Błędne dane logowania."
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autoryzacja"
                ],
                "summary": "Wyloguj użytkownika",
                "description": "Usuwa aktualny token Sanctum.",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Pomyślnie wylogowano."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invoices": {
            "get": {
                "tags": [
                    "Faktury"
                ],
                "summary": "Pobierz listę wszystkich zapisanych faktur",
                "description": "Zwraca listę wszystkich faktur (w tym tych o statusie pending/processing) wraz z ich kontrahentami, pozycjami i płatnościami. Wymaga autoryzacji.",
                "operationId": "getInvoicesList",
                "responses": {
                    "200": {
                        "description": "Pomyślnie pobrano listę faktur."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invoices/{id}": {
            "get": {
                "tags": [
                    "Faktury"
                ],
                "summary": "Pobierz szczegóły oraz status konkretnej faktury",
                "description": "Zwraca pełne dane szczegółowe oraz aktualny status ('pending', 'processing', 'completed', 'failed') wybranej faktury. Służy do odpytywania (pollingu) o stan zadania w tle.",
                "operationId": "getInvoiceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identyfikator ID faktury",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pomyślnie pobrano szczegóły faktury."
                    },
                    "404": {
                        "description": "Faktura o podanym ID nie została znaleziona."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Faktury"
                ],
                "summary": "Usuń fakturę z bazy danych",
                "description": "Usuwa wybraną fakturę oraz jej powiązany plik fizyczny z dysku.",
                "operationId": "deleteInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identyfikator ID faktury do usunięcia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Faktura pomyślnie usunięta."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invoices/upload": {
            "post": {
                "tags": [
                    "Faktury"
                ],
                "summary": "Prześlij plik PDF, JPG lub PNG i zleć przetwarzanie w tle",
                "description": "Zapisuje przesłany plik (PDF/JPG/JPEG/PNG) na dysku, tworzy rekord faktury ze statusem 'pending' i wrzuca zadanie do kolejki systemowej. Zwraca status 202 Accepted.",
                "operationId": "uploadInvoice",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "invoice_file"
                                ],
                                "properties": {
                                    "invoice_file": {
                                        "description": "Plik faktury w formacie PDF, JPG lub PNG (maks. 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Plik przesłany pomyślnie, zadanie trafiło do kolejki w tle."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Wklej tutaj swój wygenerowany token API (sam ciąg znaków, bez słowa 'Bearer')",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autoryzacja",
            "description": "Autoryzacja"
        },
        {
            "name": "Faktury",
            "description": "Faktury"
        }
    ]
}